requirements.txt.txt
streamlit
pandas
sentence-transformers
PyMuPDF
python-docx
docx2pdf
pypandoc
smtplib
email